product

+++ feedback/src/main/java/com/atos/feedback/model/Role.java
@@ -17,7 +17,7 @@ public class Role implements Serializable {
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	@Column(name="role_id")
-	private int roleId;
+	private Long roleId;
 
 	private String role;
 
@@ -28,11 +28,11 @@ public class Role implements Serializable {
 	public Role() {
 	}
 
-	public int getRoleId() {
+	public Long getRoleId() {
 		return this.roleId;
 	}
 
-	public void setRoleId(int roleId) {
+	public void setRoleId(Long roleId) {
 		this.roleId = roleId;
 	}
